Created PowerShell Script To Load Files To Azure
This is a parameterized PowerShell Script that is used to upload files to an Azure storage account. This can be used for multiple different storage accounts across multiple different azure tenants.

I have also created another script that calls the parameterized script, passing in the specific values to upload files to the intended Azure tenant and storage account.  This script access Azure key vault to be able to securely reference secret keys.